requirements.txt
pandas
requests
fake_useragent
pyspark
python-kafka
airflow
